requirement.txt
attrs==21.2.0
certifi==2021.10.8
iniconfig==1.1.1
jedi==0.18.0
packaging==21.0
parso==0.8.2
pluggy==1.0.0
pudb==2021.1
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-pudb==0.7.0
pytest-timeout==2.0.1
PyYAML==6.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
timeout==0.1.2
toml==0.10.2
urwid==2.1.2
